% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Dump_From_GeoJson}
\alias{Dump_From_GeoJson}
\title{returns a json-dump from a geojson file}
\usage{
Dump_From_GeoJson(url_file)
}
\arguments{
\item{url_file}{either a string specifying the input path to a file OR a valid url (beginning with 'http..') pointing to a geojson object}
}
\value{
a character string (json dump)
}
\description{
returns a json-dump from a geojson file
}
\examples{

\dontrun{

library(geojsonR)

res = Dump_From_GeoJson("/myfolder/point.geojson")
}

}
