% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Features_2Collection}
\alias{Features_2Collection}
\title{creates a FeatureCollection dump from multiple Feature geojson objects}
\usage{
Features_2Collection(Features_files_vec, bbox_vec = NULL)
}
\arguments{
\item{Features_files_vec}{a character vector specifying paths to files (Feature geojson objects)}

\item{bbox_vec}{either NULL or a numeric vector}
}
\value{
a FeatureCollection dump
}
\description{
creates a FeatureCollection dump from multiple Feature geojson objects
}
\examples{

library(geojsonR)

# Do not run

# vec_files = c("/myfolder/Feature1.geojson", "/myfolder/Feature2.geojson",
#               "/myfolder/Feature3.geojson", "/myfolder/Feature4.geojson",
#               "/myfolder/Feature5.geojson")

# res = Features_2Collection(vec_files, bbox_vec = NULL)

}

