% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shiny_from_JSON}
\alias{shiny_from_JSON}
\title{secondary function for shiny Applications}
\usage{
shiny_from_JSON(input_file)
}
\arguments{
\item{input_file}{a character string specifying a path to a file}
}
\value{
a (nested) list
}
\description{
secondary function for shiny Applications
}
\details{
This function is meant for \emph{shiny Applications}. To read a GeoJson file use either the \emph{FROM_GeoJson} or \emph{FROM_GeoJson_Schema} function.
}
